Minor compatibility fixes: Make gcc-4.3.1 and glibc-2.8 more happy.
git-svn-id: https://svn.gentooexperimental.org/eix/trunk@692 3c3e6e0f-8f03-0410-a2c1-9066eb21f691

+++ src/cache/cdb/cdb.cc
@@ -28,7 +28,7 @@ class Cdb {
 		bool is_ready;
 
 		uint32_t *cdb_data;
-		size_t cdb_data_size;
+		off_t cdb_data_size;
 		uint32_t *cdb_records_end;
 		uint32_t *current;
